requirements.txt
streamlit>=1.25.0
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
